Add Copyright/License boilerplate to all files
Side effects:
* Refactor file structure to group source code in `/src/` dir
* Revise notice language in LICENSE.md
* Streamline code of conduct language in CONTRIBUTING.md
