+++ LICENSE.md
@@ -1,6 +1,20 @@
-(c) Austin Guest 2015.
+NOTICE
+======
 
-This is Free Software Licensed under the GPLv3 "CopyLeft" License. To learn more about GPL, [visit](https://www.gnu.org/copyleft/gpl.html)
+Copyright (c) 2015-present, Total Location Test Paragraph.
+All rights reserved.
+
+The program in this repository is free software:
+you can redistribute it and/or modify it under the terms of
+the GNU General Public License (GPL), either version 3
+of the License, or (at your option) any later version.
+
+The program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. For more details,
+see the full license below.
+
+You can also view the GPL at <http://www.gnu.org/licenses/gpl-3.0.en.html>
 
 GNU GENERAL PUBLIC LICENSE
 ==========================
