+++ package.json
@@ -26,8 +26,8 @@
     "node-inspector": "^0.11.2"
   },
   "scripts": {
-    "test": "mocha --recursive --compilers js:babel/register",
-    "debug-test": "node-inspector & mocha --compilers js:babel/register ./test/parse.spec.js --debug-brk"
+    "test": "mocha ./src/test/**/*.spec.js --compilers js:babel-register",
+    "debug-test": "node-inspector & mocha --compilers js:babel-register ./test/parse.spec.js --debug-brk"
   },
   "repository": {
     "type": "git",
