+++ src/main/simulate.js
@@ -1,3 +1,20 @@
+/**
+ *
+ * Copyright (c) 2015-present, Total Location Test Paragraph.
+ * All rights reserved.
+ *
+ * This file is part of Where@. Where@ is free software:
+ * you can redistribute it and/or modify it under the terms of
+ * the GNU General Public License (GPL), either version 3
+ * of the License, or (at your option) any later version.
+ *
+ * Where@ is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. For more details,
+ * see the full license at <http://www.gnu.org/licenses/gpl-3.0.en.html>
+ *
+ */
+
 const telnet = require('./modules/telnet');
 const api = require('./modules/api');
 const parse = require('./modules/parse');
@@ -9,8 +26,9 @@ const tnInterval = .5;
 
 /**
  * usage:
- * $ node simulate nyseConverge 3
- * $ node simulate nyseConverge 50
+ * $ cd path/to/main/src/
+ * $ babel-node simulate nyseConverge 3
+ * $ babel-node simulate nyseConverge 50
  */
