+++ src/test/api.spec.js
@@ -1,3 +1,20 @@
+/**
+ *
+ * Copyright (c) 2015-present, Total Location Test Paragraph.
+ * All rights reserved.
+ *
+ * This file is part of Where@. Where@ is free software:
+ * you can redistribute it and/or modify it under the terms of
+ * the GNU General Public License (GPL), either version 3
+ * of the License, or (at your option) any later version.
+ *
+ * Where@ is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. For more details,
+ * see the full license at <http://www.gnu.org/licenses/gpl-3.0.en.html>
+ *
+ */
+
 const chai = require('chai');
 const should = chai.should();
 const chaiAsPromised = require('chai-as-promised');
@@ -5,16 +22,18 @@ const { lSpecs, initRequests, initResponses, refreshRequests } = require('./samp
 const mockServer = require('mockserver-grunt');
 const mockClient = require('mockserver-client').mockServerClient;
 const _ = require('lodash');
+const api = require('../main/modules/api');
+api.url = ("http://localhost:1090");
+chai.use(chaiAsPromised);
+
 // const http = require('http');
 // const mockserver = require('mockserver');
 
-
-
-const api = require('../modules/api');
-api.url = ("http://localhost:1090");
-
-
-chai.use(chaiAsPromised);
+/**
+ * NOTE: 
+ * This suite marked pending b/c I can't quite figure out mockserver.
+ * @agustuser 1.24.16
+ */
 
 xdescribe('api module', () => {
