+++ src/test/apiMocks/locations/init/POST--{ "id": "75782cd4-1a42-4af1-9130-05c63b2aa9fa", "lat": 40.704715, "lon": -74.013685, "time": 2505606400000 }.mock
