+++ src/test/apiMocks/locations/init/POST--{ "id": "75782cd4-1a42-4af1-9130-05c63b2aa9fb", "lat": 40.703084, "lon": -74.010126, "time": 2505606400000 }.mock
