+++ src/test/parse.spec.js
@@ -1,6 +1,23 @@
+/**
+ *
+ * Copyright (c) 2015-present, Total Location Test Paragraph.
+ * All rights reserved.
+ *
+ * This file is part of Where@. Where@ is free software:
+ * you can redistribute it and/or modify it under the terms of
+ * the GNU General Public License (GPL), either version 3
+ * of the License, or (at your option) any later version.
+ *
+ * Where@ is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. For more details,
+ * see the full license at <http://www.gnu.org/licenses/gpl-3.0.en.html>
+ *
+ */
+
 const chai = require('chai');
 const should = chai.should();
-const parse = require('../modules/parse');
+const parse = require('../main/modules/parse');
 const { lSpecs, updateRequests, initRequests, refreshRequests, ids } = require('./samples');
 const { max, floor } = require('lodash');
 const { abs } = Math;
